Added rudimentary unit testing support for repository. Moved startup logic to own class. Add some new business methods in ImageService. Add safer logic and error messages to ShellHandler.

+++ src/main/java/com/joshuaharwood/limelight/server/service/shell/ShellHandler.java
@@ -30,9 +30,47 @@ public class ShellHandler {
     }
 
     @ShellMethod("Delete an Image entity from the database")
-    public String deleteImageById(Long id) {
-        imageService.deleteImage(id);
-        return imageService.getImageStats();
+    public String deleteImage(Long id) {
+        if (imageService.imageExistsById(id)) {
+            imageService.deleteImage(id);
+            return String.format("Deleted image id %d", id);
+        } else {
+            return String.format("Image not found for id %d", id);
+        }
+    }
+
+    @ShellMethod("Delete an Image entity's associated content")
+    public String deleteImageContent(Long id) {
+        if (imageService.imageExistsById(id)) {
+            imageService.deleteImageContent(id);
+            return String.format("Deleted image content for id %d", id);
+        } else {
+            return String.format("Image not found for id %d", id);
+        }
+    }
+
+    @ShellMethod("Create an image entity")
+    public String createImage() {
+        return String.format("Created image with id %d",
+                imageService.createImage().getId());
+    }
+
+    @ShellMethod("Create multiple image entities")
+    public String createImages(Long num) {
+        for (int i = 0; i < num; i++) {
+            imageService.createImage();
+        }
+
+        return String.format("Created %d image(s)", num);
+    }
+
+    @ShellMethod("Check if image has content")
+    public String checkIfHasContent(Long id) {
+        if (imageService.imageExistsById(id)) {
+            return String.valueOf(imageService.imageHasContent(id));
+        } else {
+            return String.format("Image not found for id %d", id);
+        }
     }
 
     @PostConstruct
